apps/backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
async-lru==2.0.5
beanie==1.25.0
boto3==1.39.3
botocore==1.39.3
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.3.0
Faker==37.4.0
fastapi==0.115.0
fastapi-cli==0.0.7
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
lazy-model==0.2.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
motor==3.4.0
orjson==3.10.18
passlib==1.7.4
pycparser==2.22
pydantic==2.7.1
pydantic-extra-types==2.10.5
pydantic-settings==2.2.1
pydantic_core==2.18.2
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.4
rich==14.0.0
rich-toolkit==0.14.8
s3transfer==0.13.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.38.6
stripe==9.4.0
toml==0.10.2
typer==0.16.0
typing_extensions==4.14.0
tzdata==2025.2
ujson==5.10.0
urllib3==2.5.0
uvicorn==0.29.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
